修改数字字段的长度

+++ backends/apps/qa/models.py
@@ -21,7 +21,7 @@ class QaHead(models.Model):
     ftestusr = models.CharField(verbose_name='测试者', max_length=24, null=True, blank=True)
     fconfirmdte = models.DateField(verbose_name='确认日期', null=True, blank=True)
     fconfirmusr = models.CharField(verbose_name='确认者', max_length=24, null=True, blank=True)
-    fttlcodelines = models.SmallIntegerField(verbose_name='影响总行数', default=0)
+    fttlcodelines = models.IntegerField(verbose_name='影响总行数', default=0)
     fmodifiedlines = models.SmallIntegerField(verbose_name='修改行数', default=0)
     fcomplexity = models.DecimalField(verbose_name='复杂度', max_digits=2, decimal_places=1, null=True, blank=True)
     fnote = models.TextField(verbose_name='备注', null=True, blank=True)
